Emergency push

+++ database/migrations/2023_12_11_181920_create_files_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
 
             $table->string('pathname');
             $table->string('mime');
-            $table->float('size');
+            $table->integer('size');
 
             $table->timestamps();
             $table->softDeletes();
